Added NavMesh and Basic Movement
Learned about “Align with View” option from the gameObject tab, which allows you to move an object to wherever your scene view currently is. This is super handy for moving the camera. Added a navigation mesh, allowing the player to move to places in the world. Completed the challenge on my own! Getting access to a component added on to an object through script, and researched the documentation to fin which namespace was needed.